Restructure compiler-plugin module into: backend, k2, common and cli (#149)
