+++ compiler-plugin/compiler-plugin-backend/build.gradle.kts
@@ -1,0 +1,19 @@
+/*
+ * Copyright 2023-2024 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+import org.jetbrains.kotlin.gradle.dsl.ExplicitApiMode
+
+plugins {
+    alias(libs.plugins.conventions.jvm)
+    alias(libs.plugins.compiler.specific.module)
+}
+
+kotlin {
+    explicitApi = ExplicitApiMode.Disabled
+}
+
+dependencies {
+    compileOnly(libs.kotlin.compiler.embeddable)
+    implementation(projects.compilerPluginCommon)
+}
