+++ compiler-plugin/compiler-plugin-backend/src/main/core/kotlinx/rpc/codegen/RPCIrPlugin.kt
@@ -1,0 +1,20 @@
+/*
+ * Copyright 2023-2024 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package kotlinx.rpc.codegen
+
+import kotlinx.rpc.codegen.extension.RPCIrExtension
+import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
+import org.jetbrains.kotlin.cli.common.CLIConfigurationKeys
+import org.jetbrains.kotlin.cli.common.messages.MessageCollector
+import org.jetbrains.kotlin.config.CompilerConfiguration
+
+object RPCIrPlugin {
+    fun provideExtension(configuration: CompilerConfiguration): IrGenerationExtension {
+        val logger = configuration.get(CLIConfigurationKeys.MESSAGE_COLLECTOR_KEY, MessageCollector.NONE)
+        val versionSpecificApi = VersionSpecificApi.INSTANCE
+
+        return RPCIrExtension(logger, versionSpecificApi)
+    }
+}
