+++ compiler-plugin/compiler-plugin-cli/src/main/core/Stub.kt
@@ -1,0 +1,7 @@
+/*
+ * Copyright 2023-2024 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+@file:Suppress("detekt.MissingPackageDeclaration")
+
+// do not delete
