+++ compiler-plugin/compiler-plugin-cli/src/main/resources/latest/META-INF/services/org.jetbrains.kotlin.compiler.plugin.CompilerPluginRegistrar
