+++ compiler-plugin/compiler-plugin-cli/src/main/resources/v_1_7/META-INF/services/org.jetbrains.kotlin.compiler.plugin.ComponentRegistrar
