+++ compiler-plugin/compiler-plugin-cli/src/main/resources/v_1_7_2/META-INF/services/org.jetbrains.kotlin.compiler.plugin.ComponentRegistrar
