+++ compiler-plugin/compiler-plugin-cli/src/main/resources/v_1_8/META-INF/services/org.jetbrains.kotlin.compiler.plugin.CommandLineProcessor
