+++ gradle-conventions/conventions-utils/src/main/kotlin/conventions-utils.gradle.kts
@@ -1,0 +1,5 @@
+/*
+ * Copyright 2023-2024 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+// do not delete. This plugin includes util functions in the buildscript classpath
